- Code review, cleanup and documentation
