+++ src/bin/printhor/control/motion/mod.rs
@@ -1,0 +1,3 @@
+pub mod profile;
+
+pub use profile::*;
