--- src/bin/printhor/control/motion_planning/mod.rs
@@ -1,4 +1,0 @@
-mod interpolators;
-pub mod plan;
-
-pub use plan::*;
