+++ src/bin/printhor/control/task_stepper.rs
@@ -11,41 +11,87 @@
 //!     - Consume a micro-segment until iterator is exhausted
 //!
 //! TODO: This is a work still in progress
-//!
-//! TODO: Refactor pending
-//!
-use crate::control::motion_planning::SCurveMotionProfile;
-#[allow(unused_imports)]
+use crate::control::motion::SCurveMotionProfile;
 use crate::hwa;
 use crate::math;
-#[allow(unused)]
-use crate::math::{Real, ONE_HUNDRED, ONE_MILLION, ONE_THOUSAND};
-#[allow(unused_imports)]
+use crate::math::Real;
 use crate::tgeo::{CoordSel, TVector};
-use core::cell::RefCell;
-use core::future::{poll_fn, Future};
-use core::task::{Context, Poll};
-use embassy_sync::waitqueue::WakerRegistration;
+use num_traits::ToPrimitive;
 use hwa::controllers::motion::SegmentIterator;
 use printhor_hwa_common::{DeferAction, DeferEvent};
 use printhor_hwa_common::{EventBusRef, StepperChannel};
-#[allow(unused)]
 use printhor_hwa_common::{EventFlags, EventStatus};
-
-use super::motion_timing::*;
+use crate::hwa::controllers::LinearMicrosegmentStepInterpolator;
+use crate::hwa::controllers::motion::STEP_DRIVER;
 
 const DO_NOTHING: bool = false;
 
 /// Inactivity Timeout until steppers are disabled
 const STEPPER_INACTIVITY_TIMEOUT: embassy_time::Duration = embassy_time::Duration::from_secs(5);
 
+
+///
+/// This constant defines the micro-segment period for the stepper planner in microseconds.
+///
+/// The period represents the duration of each micro-segment in the stepper motor control algorithm, 
+/// calculated based on the `STEPPER_PLANNER_MICROSEGMENT_FREQUENCY`. It is used to break down motion 
+/// profiles into smaller time intervals to achieve precise control of stepper motors.
+///
+/// # Calculation
+/// This constant value is computed by dividing 1,000,000 microseconds (which is equivalent to 1 second)
+/// by the `STEPPER_PLANNER_MICROSEGMENT_FREQUENCY`.
+///
+/// # Usage
+/// The constant is used in the stepper task to determine the sampling period for evaluating motion 
+/// profiles and generating stepping pulses.
+///
+/// # Note
+/// Ensure that the `STEPPER_PLANNER_MICROSEGMENT_FREQUENCY` is correctly set to match the desired motion 
+/// control frequency.
+///
 const STEPPER_PLANNER_MICROSEGMENT_PERIOD_US: u32 =
     1_000_000 / hwa::STEPPER_PLANNER_MICROSEGMENT_FREQUENCY;
-const STEPPER_PLANNER_CLOCK_PERIOD_US: u32 = 1_000_000 / hwa::STEPPER_PLANNER_CLOCK_FREQUENCY;
 
-/***
-This task feeds watchdog to ensure no reset happen due high CPU starvation when feed rate is very high
- */
+/**
+This constant defines the micro-segment period for the stepper planner in microseconds.
+
+It represents the duration of each micro-segment in the stepper motor control algorithm, which is 
+calculated based on the `STEPPER_PLANNER_MICROSEGMENT_FREQUENCY`. A micro-segment is a small 
+time interval used to break down motion profiles for precise control of stepper motors.
+
+# Calculation
+The constant value is computed by dividing 1,000,000 microseconds (equivalent to 1 second) by 
+the `STEPPER_PLANNER_MICROSEGMENT_FREQUENCY`.
+
+# Usage
+This constant is used in the stepper task to determine the sampling period for evaluating 
+motion profiles and generating stepping pulses.
+
+# Note
+Ensure the `STEPPER_PLANNER_MICROSEGMENT_FREQUENCY` is set correctly to match the desired motion 
+control frequency.
+*/
+pub const STEPPER_PLANNER_CLOCK_PERIOD_US: u32 = 1_000_000 / hwa::STEPPER_PLANNER_CLOCK_FREQUENCY;
+
+/// This asynchronous task manages the steps of the stepper motors 
+/// based on motion segments from the motion queue.
+///
+/// # Arguments
+///
+/// * `event_bus` - Reference to the event bus for handling events.
+/// * `motion_planner` - Reference to the motion planner controller.
+/// * `_watchdog` - Reference to the watchdog controller.
+///
+/// # Behavior
+///
+/// - Waits for the `ATX_ON` signal to proceed.
+/// - Disables steppers upon inactivity or `SYS_ALARM`.
+/// - Retrieves and processes motion segments to drive the stepper motors.
+/// - Handles homing procedure if no segments are available.
+/// - This task feeds watchdog to ensure no reset happen due high CPU starvation when feed rate is very high
+///
+/// This function runs indefinitely and controls the stepper motors
+/// according to the motion planning profile.
 #[embassy_executor::task]
 pub async fn task_stepper(
     event_bus: EventBusRef,
@@ -62,7 +108,7 @@ pub async fn task_stepper(
 
     motion_planner.start(&event_bus).await;
 
-    TM.setup(motion_planner.motion_driver());
+    STEP_DRIVER.setup(motion_planner.motion_driver());
 
     let mut s = event_bus.subscriber().await;
 
@@ -91,13 +137,13 @@ pub async fn task_stepper(
             hwa::info!("task_stepper waiting for ATX_ON");
             // For safely, disable steppers
             // TODO: Park
-            TM.flush().await;
+            STEP_DRIVER.flush().await;
             motion_planner
                 .motion_driver
                 .lock()
                 .await
                 .disable_steppers(StepperChannel::all());
-            TM.reset();
+            STEP_DRIVER.reset();
             steppers_off = true;
 
             if s.ft_wait_until(EventFlags::ATX_ON).await.is_err() {
@@ -110,13 +156,13 @@ pub async fn task_stepper(
         if wait_for_sysalarm || s.get_status().await.contains(EventFlags::SYS_ALARM) {
             hwa::warn!("task stepper waiting for SYS_ALARM release");
             // TODO: Park
-            TM.flush().await;
+            STEP_DRIVER.flush().await;
             motion_planner
                 .motion_driver
                 .lock()
                 .await
                 .disable_steppers(StepperChannel::all());
-            TM.reset();
+            STEP_DRIVER.reset();
             steppers_off = true;
             if s.ft_wait_while(EventFlags::SYS_ALARM).await.is_err() {
                 panic!("Unexpected situation");
@@ -126,20 +172,20 @@ pub async fn task_stepper(
             STEPPER_INACTIVITY_TIMEOUT,
             motion_planner.get_current_segment_data(&event_bus),
         )
-        .await
+            .await
         {
             // Timeout
             Err(_) => {
                 hwa::trace!("stepper_task timeout");
                 if !steppers_off {
                     hwa::info!("Timeout. Powering steppers off");
-                    TM.flush().await;
+                    STEP_DRIVER.flush().await;
                     motion_planner
                         .motion_driver
                         .lock()
                         .await
                         .disable_steppers(StepperChannel::all());
-                    TM.reset();
+                    STEP_DRIVER.reset();
                     steppers_off = true;
                 }
             }
@@ -309,15 +355,15 @@ pub async fn task_stepper(
 
                                 #[cfg(feature = "verbose-timings")]
                                 let t1 = embassy_time::Instant::now();
-                                TM.push(
+                                STEP_DRIVER.push(
                                     microsegment_interpolator.state().clone(),
                                     stepper_enable_flags,
                                     stepper_dir_fwd_flags,
                                 )
-                                .await;
+                                    .await;
                                 cfg_if::cfg_if! {
                                     if #[cfg(feature="assert-motion")] {
-                                        steps_advanced += TM.flush().await;
+                                        steps_advanced += STEP_DRIVER.flush().await;
                                     }
                                 }
                                 #[cfg(feature = "verbose-timings")]
@@ -428,7 +474,7 @@ pub async fn task_stepper(
 
                         cfg_if::cfg_if! {
                             if #[cfg(feature="verbose-timings")] {
-                                use crate::control::motion_planning::plan::MotionProfile;
+                                use crate::control::motion::profile::MotionProfile;
 
                                 //global_timer = embassy_time::Instant::now();
                                 hwa::info!("\t[v_0 = {}, v_lim = {}, v_1 = {}, t = {} d = {} mm = {} stp = {}]; {} moves left",
@@ -457,13 +503,13 @@ pub async fn task_stepper(
             Ok(None) => {
                 hwa::debug!("Homing init");
 
-                TM.flush().await;
+                STEP_DRIVER.flush().await;
                 motion_planner
                     .motion_driver
                     .lock()
                     .await
                     .enable_steppers(StepperChannel::all());
-                TM.reset();
+                STEP_DRIVER.reset();
 
                 if steppers_off {
                     hwa::info!("\tPowering steppers on");
@@ -487,492 +533,4 @@ pub async fn task_stepper(
             }
         }
     }
-}
-
-pub struct LinearMicrosegmentStepInterpolator {
-    vdir_abs: TVector<Real>,
-    usteps_per_mm: TVector<Real>,
-
-    /// Number of discrete steps along vector already advanced
-    usteps_advanced: TVector<u32>,
-
-    /// The number of discrete steps by axis already advanced
-    axis_steps_advanced_precise: TVector<Real>,
-
-    /// The number of discrete steps by axis to advance
-    axis_steps_to_advance_precise: TVector<Real>,
-
-    multi_timer: MultiTimer,
-    #[cfg(any(test, feature = "assert-motion"))]
-    delta: TVector<u32>,
-}
-
-impl LinearMicrosegmentStepInterpolator {
-    pub fn new(vdir_abs: TVector<Real>, distance: Real, usteps_per_mm: TVector<Real>) -> Self {
-        Self {
-            vdir_abs,
-            usteps_per_mm,
-            usteps_advanced: TVector::zero(),
-            axis_steps_advanced_precise: TVector::zero(),
-            axis_steps_to_advance_precise: (vdir_abs * distance * usteps_per_mm).floor(),
-            multi_timer: MultiTimer::new(),
-            #[cfg(any(test, feature = "assert-motion"))]
-            delta: TVector::zero(),
-        }
-    }
-
-    pub fn advance_to(&mut self, estimated_position: Real, width: Real) -> bool {
-        let axial_pos: TVector<Real> = self.vdir_abs * estimated_position;
-        let step_pos: TVector<Real> = axial_pos * self.usteps_per_mm;
-        let steps_to_advance = (step_pos - self.axis_steps_advanced_precise)
-            .clamp_min(TVector::zero())
-            .clamp(self.axis_steps_to_advance_precise);
-        let steps_to_advance_precise_1: TVector<Real> = steps_to_advance;
-        let steps_to_advance_precise = steps_to_advance_precise_1.floor();
-        self.axis_steps_advanced_precise += steps_to_advance_precise;
-
-        #[cfg(any(test, feature = "assert-motion"))]
-        {
-            self.delta = steps_to_advance_precise.map_all(|c| {
-                c.unwrap_or(math::ZERO)
-                    .to_i32()
-                    .and_then(|i| Some(i as u32))
-            });
-        }
-
-        let _can_advance_more = self
-            .axis_steps_advanced_precise
-            .bounded_by(&self.axis_steps_to_advance_precise);
-
-        // Centering formula:
-        // x = width + num_steps / num_steps
-
-        let numerator = width.to_i32().unwrap_or(0) as u32;
-        if numerator == 0 {
-            panic!("bad advance");
-        }
-        let tick_period_by_axis = steps_to_advance_precise.map_all(|c| {
-            c.unwrap_or(math::ZERO).to_i32().and_then(|i| {
-                if i > 0 {
-                    Some(numerator / (i as u32))
-                } else {
-                    None
-                }
-            })
-        });
-
-        let step_increment =
-            steps_to_advance_precise.map_coords(|cv| cv.to_i32().and_then(|c| Some(c as u32)));
-
-        self.usteps_advanced += step_increment;
-        self.multi_timer.set_width(numerator);
-        self.multi_timer.set_max_count(&step_increment);
-
-        #[cfg(test)]
-        {
-            std::println!("...");
-            std::println!(
-                "step_pos: {} steps_to_advance : {}",
-                step_pos,
-                steps_to_advance
-            );
-            std::println!(
-                "delta: {} steps_to_advance_precise : {}",
-                self.delta,
-                steps_to_advance_precise
-            );
-            std::println!(
-                "width: {} tick period by axis: {}",
-                numerator,
-                tick_period_by_axis
-            );
-        }
-
-        #[cfg(feature = "with-x-axis")]
-        self.multi_timer
-            .set_channel_ticks(StepperChannel::X, tick_period_by_axis.x);
-        #[cfg(feature = "with-y-axis")]
-        self.multi_timer
-            .set_channel_ticks(StepperChannel::Y, tick_period_by_axis.y);
-        #[cfg(feature = "with-z-axis")]
-        self.multi_timer
-            .set_channel_ticks(StepperChannel::Z, tick_period_by_axis.z);
-        #[cfg(feature = "with-e-axis")]
-        self.multi_timer
-            .set_channel_ticks(StepperChannel::E, tick_period_by_axis.e);
-        //can_advance_more
-        true
-    }
-
-    pub fn state(&self) -> &MultiTimer {
-        &self.multi_timer
-    }
-
-    #[inline]
-    pub fn advanced_steps(&self) -> TVector<u32> {
-        self.usteps_advanced
-    }
-
-    #[allow(unused)]
-    pub fn advanced_mm(&self) -> TVector<Real> {
-        self.advanced_steps()
-            .map_coords(|c| Some(Real::from_lit(c.into(), 0)))
-            / self.usteps_per_mm
-    }
-
-    #[allow(unused)]
-    pub fn width(&self) -> u32 {
-        self.multi_timer.width()
-    }
-}
-
-#[cfg(test)]
-pub mod test {
-
-    #[cfg(feature = "wip-tests")]
-    #[test]
-    fn interpolation_test_1() {
-        use crate::control::task_stepper::LinearMicrosegmentStepInterpolator;
-        use crate::math;
-        use crate::math::Real;
-        use crate::tgeo::TVector;
-        let mut lin = LinearMicrosegmentStepInterpolator::new(
-            TVector::from_coords(Some(math::ONE), None, None, None),
-            Real::from_f32(100.0),
-            TVector::from_coords(Some(Real::from_f32(160.0)), None, None, None),
-        );
-
-        lin.advance_to(Real::from_f32(100.0), Real::from_f32(100.0));
-
-        std::println!("xx");
-    }
-}
-
-use crate::hwa::drivers::motion_driver::MotionDriverRef;
-use critical_section::Mutex as CsMutex;
-use num_traits::ToPrimitive;
-
-const TIMER_QUEUE_SIZE: usize = 4;
-
-pub struct SoftTimerDriver {
-    current: StepPlanner,
-    pub state: State,
-    tick_count: u32,
-    queue: [Option<StepPlanner>; TIMER_QUEUE_SIZE],
-    head: u8,
-    tail: u8,
-    num_queued: u8,
-    drv: Option<MotionDriverRef>,
-    waker: WakerRegistration,
-    current_stepper_enable_flags: StepperChannel,
-    current_stepper_dir_fwd_flags: StepperChannel,
-    #[cfg(any(feature = "assert-motion", test))]
-    pub pulses: [u32; 4],
-}
-
-#[derive(PartialEq)]
-pub enum State {
-    Idle,
-    Duty,
-}
-
-impl SoftTimerDriver {
-    fn try_consume(&mut self) -> bool {
-        if self.num_queued > 0 {
-            if let Some(timer) = self.queue[self.head as usize].take() {
-                self.current = timer;
-                self.tick_count = 0;
-                match &self.drv {
-                    None => {
-                        unreachable!("Driver instance not set")
-                    }
-                    Some(_ref) => match _ref.try_lock() {
-                        Ok(mut _drv) => {
-                            #[cfg(feature = "debug-signals")]
-                            _drv.pins.y_dir_pin.set_high();
-                            if self.current_stepper_enable_flags
-                                != self.current.stepper_enable_flags
-                            {
-                                _drv.pins.enable(self.current.stepper_enable_flags);
-                                self.current_stepper_enable_flags =
-                                    self.current.stepper_enable_flags;
-                            }
-                            if self.current_stepper_dir_fwd_flags
-                                != self.current.stepper_dir_fwd_flags
-                            {
-                                _drv.pins
-                                    .set_forward_direction(self.current.stepper_dir_fwd_flags);
-                                self.current_stepper_dir_fwd_flags =
-                                    self.current.stepper_dir_fwd_flags;
-                            }
-                        }
-                        Err(_) => {
-                            unreachable!("unable to lock")
-                        }
-                    },
-                }
-                hwa::trace!("u-segment dequeued at {}", self.head);
-            } else {
-                unreachable!("Unable to deque!!!")
-            }
-            self.state = State::Duty;
-            self.head += 1;
-            if self.head as usize == self.queue.len() {
-                self.head = 0;
-            }
-            self.num_queued -= 1;
-            true
-        } else {
-            self.state = State::Idle;
-            // Quickly return as there is no work to do
-            false
-        }
-    }
-
-    pub fn notify(&mut self) {
-        if (self.num_queued as usize) < TIMER_QUEUE_SIZE {
-            self.waker.wake();
-        }
-    }
-
-    pub fn on_interrupt(&mut self) {
-        if self.state == State::Idle {
-            if !self.try_consume() {
-                self.notify();
-                return;
-            }
-        }
-        self.tick_count += STEPPER_PLANNER_CLOCK_PERIOD_US;
-
-        // In this point, state is either Duty or something were dequeued
-        match self.current.next(STEPPER_PLANNER_CLOCK_PERIOD_US) {
-            None => {
-                panic!("Unexpected state");
-            }
-            Some(_ch) => {
-                self.apply(_ch);
-            }
-        }
-
-        if self.tick_count >= self.current.interval_width {
-            hwa::trace!("u-segment completed in {} us", self.tick_count);
-            if self.try_consume() {}
-            #[cfg(feature = "debug-signals")]
-            match &self.drv {
-                None => {
-                    unreachable!("Driver instance not set")
-                }
-                Some(_ref) => match _ref.try_lock() {
-                    Ok(mut _drv) => {
-                        _drv.pins.y_dir_pin.set_low();
-                    }
-                    Err(_) => {
-                        unreachable!("unable to lock")
-                    }
-                },
-            }
-            self.notify();
-        }
-    }
-
-    fn apply(&mut self, _channel: StepperChannel) -> bool {
-        if _channel.is_empty() {
-            true
-        } else {
-            match &self.drv {
-                None => false,
-                Some(_ref) => match _ref.try_lock() {
-                    Ok(mut _drv) => {
-                        cfg_if::cfg_if! {
-                            if #[cfg(feature = "pulsed")] {
-                                _drv.step_pin_high(_channel);
-                                s_block_for(STEPPER_PULSE_WIDTH_US);
-                                _drv.step_pin_high(_channel);
-                            }
-                            else {
-                                _drv.step_toggle(_channel);
-                            }
-                        }
-                        #[cfg(feature = "assert-motion")]
-                        {
-                            if _channel.contains(StepperChannel::X) {
-                                self.pulses[0] += 1;
-                            }
-                            if _channel.contains(StepperChannel::Y) {
-                                self.pulses[1] += 1;
-                            }
-                            if _channel.contains(StepperChannel::Z) {
-                                self.pulses[2] += 1;
-                            }
-                            if _channel.contains(StepperChannel::E) {
-                                self.pulses[3] += 1;
-                            }
-                        }
-                        true
-                    }
-                    Err(_) => false,
-                },
-            }
-        }
-    }
-}
-
-pub struct SoftTimer(pub CsMutex<RefCell<SoftTimerDriver>>);
-
-cfg_if::cfg_if! {
-    if #[cfg(any(feature = "assert-motion", test))] {
-        pub type FlushResult = TVector<u32>;
-    }
-    else {
-        pub type FlushResult = ();
-    }
-}
-
-impl SoftTimer {
-    pub const fn new() -> Self {
-        Self(CsMutex::new(RefCell::new(SoftTimerDriver {
-            current: StepPlanner::new(),
-            state: State::Idle,
-            queue: [None; TIMER_QUEUE_SIZE],
-            head: 0,
-            tick_count: 0,
-            waker: WakerRegistration::new(),
-            num_queued: 0,
-            tail: 0,
-            drv: None,
-            current_stepper_enable_flags: StepperChannel::empty(),
-            current_stepper_dir_fwd_flags: StepperChannel::empty(),
-            #[cfg(any(feature = "assert-motion", test))]
-            pulses: [0, 0, 0, 0],
-        })))
-    }
-
-    pub fn setup(&self, _mp: MotionDriverRef) {
-        critical_section::with(|cs| {
-            let mut r = self.0.borrow_ref_mut(cs);
-            r.drv.replace(_mp);
-        })
-    }
-
-    pub fn push(
-        &self,
-        multi_timer: MultiTimer,
-        stepper_enable_flags: StepperChannel,
-        stepper_dir_fwd_flags: StepperChannel,
-    ) -> impl Future<Output = ()> + '_ {
-        poll_fn(move |cx| {
-            self.poll_push(cx, multi_timer, stepper_enable_flags, stepper_dir_fwd_flags)
-        })
-    }
-
-    fn poll_push(
-        &self,
-        cx: &mut Context<'_>,
-        multi_timer: MultiTimer,
-        stepper_enable_flags: StepperChannel,
-        stepper_dir_fwd_flags: StepperChannel,
-    ) -> Poll<()> {
-        critical_section::with(|cs| {
-            let mut r = self.0.borrow_ref_mut(cs);
-            if r.num_queued as usize >= r.queue.len() {
-                r.waker.register(cx.waker());
-                Poll::Pending
-            } else {
-                r.num_queued += 1;
-                let current_tail = r.tail;
-                hwa::trace!("u-segment queued at {}", current_tail);
-                r.queue[current_tail as usize] = Some(StepPlanner::from(
-                    multi_timer,
-                    stepper_enable_flags,
-                    stepper_dir_fwd_flags,
-                ));
-                r.tail = if (r.tail + 1) as usize == r.queue.len() {
-                    0
-                } else {
-                    r.tail + 1
-                };
-                Poll::Ready(())
-            }
-        })
-    }
-
-    #[allow(unused)]
-    pub fn flush(&self) -> impl Future<Output = FlushResult> + '_ {
-        poll_fn(move |cx| self.poll_flush(cx))
-    }
-
-    pub fn reset(&self) {
-        critical_section::with(|cs| {
-            let mut r = self.0.borrow_ref_mut(cs);
-            r.current_stepper_dir_fwd_flags = StepperChannel::UNSET;
-            r.current_stepper_dir_fwd_flags = StepperChannel::UNSET;
-        });
-    }
-
-    fn poll_flush(&self, cx: &mut Context<'_>) -> Poll<FlushResult> {
-        critical_section::with(|cs| {
-            let mut r = self.0.borrow_ref_mut(cs);
-            if r.num_queued > 0 {
-                r.waker.register(cx.waker());
-                Poll::Pending
-            } else {
-                cfg_if::cfg_if! {
-                    if #[cfg(any(feature = "assert-motion", test))] {
-                        let pulses = TVector::from_coords(
-                            Some(r.pulses[0]), Some(r.pulses[1]), Some(r.pulses[2]), Some(r.pulses[3])
-                        );
-                        hwa::trace!("Pulses: {}", pulses);
-                        r.pulses[0] = 0;
-                        r.pulses[1] = 0;
-                        r.pulses[2] = 0;
-                        r.pulses[3] = 0;
-                        Poll::Ready(pulses)
-                    }
-                    else {
-                        Poll::Ready(())
-                    }
-                }
-            }
-        })
-    }
-}
-
-static TM: SoftTimer = SoftTimer::new();
-
-#[no_mangle]
-pub extern "Rust" fn do_tick() {
-    critical_section::with(|cs| {
-        let mut counter = TM.0.borrow_ref_mut(cs);
-        #[cfg(feature = "debug-signals")]
-        match &counter.drv {
-            None => {
-                //unreachable!("Driver instance not set")
-            }
-            Some(_ref) => {
-                match _ref.try_lock() {
-                    Ok(mut _drv) => {
-                        _drv.pins.x_dir_pin.set_high();
-                    }
-                    Err(_e) => {
-                        //unreachable!("Unable to lock")
-                    }
-                }
-            }
-        }
-        counter.on_interrupt();
-        #[cfg(feature = "debug-signals")]
-        match &counter.drv {
-            None => {}
-            Some(_ref) => {
-                match _ref.try_lock() {
-                    Ok(mut _drv) => {
-                        _drv.pins.x_dir_pin.set_low();
-                    }
-                    Err(_e) => {
-                        //unreachable!("Unable to lock")
-                    }
-                }
-            }
-        }
-    });
 }
